refactor(db): address review — rename utils/ssl.py to utils/tls.py; clearer env docs
- Rename onyx/utils/ssl.py to onyx/utils/tls.py so it no longer shadows the
  stdlib `ssl` module (a foot-gun even though absolute imports keep it safe).
- Rewrite the docker-compose TLS env block as a layered guide: one-way SSL is
  the common case (require -> verify-ca -> verify-full), mTLS is the rare add-on.
