+++ deployment/docker_compose/env.template
@@ -111,17 +111,24 @@ POSTGRES_PASSWORD=password
 # POSTGRES_TCP_KEEPALIVES_IDLE=
 # POSTGRES_TCP_KEEPALIVES_INTERVAL=
 # POSTGRES_TCP_KEEPALIVES_COUNT=
-# TLS for the database connection (libpq sslmode vocabulary). Only verify-ca /
-# verify-full authenticate the server, and they REQUIRE POSTGRES_SSLROOTCERT (a
-# readable CA bundle path; use the system bundle for publicly-trusted certs).
-# Ignored when USE_IAM_AUTH=true.
+# --- Database TLS (libpq sslmode vocabulary; ignored when USE_IAM_AUTH=true) ---
+# Pick the level you need; each builds on the one above. Default (unset) = no TLS.
+#
+#   require       encrypt only — does NOT detect a MITM (server is not authenticated)
+#   verify-ca     + verify the server cert chains to POSTGRES_SSLROOTCERT
+#   verify-full   + also verify the server hostname  <- recommended for managed DBs
+#
+# verify-ca / verify-full REQUIRE POSTGRES_SSLROOTCERT (CA bundle path; use the
+# system bundle for publicly-trusted server certs).
 # POSTGRES_SSLMODE=          # disable | allow | prefer | require | verify-ca | verify-full
 # POSTGRES_SSLROOTCERT=      # e.g. /etc/ssl/certs/rds-ca-bundle.pem
-# Mutual TLS (optional): present a client certificate. Both must be set, under a
-# POSTGRES_SSLMODE that negotiates SSL (require / verify-ca / verify-full).
+#
+# Mutual TLS (optional, rarely needed): also present a CLIENT cert so the server
+# can authenticate us. Most deployments do NOT set these. Requires a POSTGRES_SSLMODE
+# that negotiates SSL (require / verify-ca / verify-full); cert and key must both be set.
 # POSTGRES_SSLCERT=          # client certificate path
 # POSTGRES_SSLKEY=           # client private key path
-# POSTGRES_SSLKEY_PASSWORD=  # optional, if the key is encrypted
+# POSTGRES_SSLKEY_PASSWORD=  # optional, only if the key file is encrypted
 # DB_READONLY_USER=
 # DB_READONLY_PASSWORD=
